requirements.txt
google-generativeai
fastapi
uvicorn
psycopg2-binary
